requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord==1.0.1
discord.py==1.4.1
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
typing-extensions==3.7.4.3
yarl==1.5.1
Pillow==7.1.2
requests==2.23.0
PyNaCl==1.3.0
youtube-dl==2020.12.9
pytube==10.0.0